Add prisma and update shuttle and announcement endpoints (#35)
This PR adds prisma. It also converts the `/shuttles` and
`/announcement` endpoints to fetch data from the database. Example data
can be added to the database using the `seed.ts` file. This also adds
the ability to filter announcements by type.

This does not introduce any authentication on the endpoints.