Add API keys (#38)
This adds support for API keys in requests. Users will send their API
keys to endpoints that require them by adding `Bearer <api-key>` to the
Authentication header in their request.

When a request needs a certain API key scope, use the
`requireAPIKeyScopes()` function defined in the `api-middleware.ts` file
in the middleware folder. This function takes an array of scopes.

You can add your own scopes in the prisma schema.