Add tests & actions workflow (#55)
This adds an actions workflow that runs the newly created tests. This
also updates the license to add the original appendix back and adds a
header to our files.

For the `/announcements` tests, I still need to figure out to ensure
that the correct type enum is being returned. This will be done in the
future if we have time.

This removes the faculty data that was being used for the seed file and
instead adds a couple to the seed file. This also renames the initial
prisma migration to `initial`.